requirements.txt
streamlit 
google-generativeai 
python-dotenv 
langchain 
PyPDF2 
chromadb 
faiss-cpu 
langchain_google_genai
langchain-community
dateparser 
validators
parsedatetime